Adds documentation and small cleanups.
